requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
eval_type_backport==0.2.0
filelock==3.16.0
frozenlist==1.4.1
idna==3.9
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.1.0
numpy==2.1.1
packaging==24.1
pillow==10.4.0
pipdeptree==2.23.4
pyarrow==17.0.0
pydantic==2.9.1
pydantic_core==2.23.3
Pygments==2.18.0
python-dotenv==1.0.1
requests==2.32.3
rich==13.8.1
shellingham==1.5.4
tabulate==0.9.0
together==1.2.12
tqdm==4.66.5
typer==0.12.5
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.11.1
